Turn Symbol class into an interface
In order to enable mixins and using a same set of features for different
branches of the Symbol class hierarchy, the Symbol class was turned into
an interface and a new class, serving as the abstract base class for real
implementations, was created (see AbstractSymbol).

A related extension is moving the DelegationSymbol class into the core and
providing a new PublicSymbol interface that embodies the meaning of the
ExternalSymbol class (which was turned into the default implementation).

TaggedSymbol, thanks to switching to Symbol as an interface, was turned
into an interface as well.