Introduce the concept of mappable symbols
A new interface, Mappable, has been inherited from AdaptationStrategy and
became the base for Symbol. This new interface declares generic adaptive
mapping using the key whose origin is deferred to the inherited classes.

The Symbol interface so far provided (almost) the same set of methods like
the new Mappable interface, but also provided the default implementation
using 'this' as the only possible key.

When the PublicSymbol interface was about to be extended to support easily
interfacing with Map instances with the defined key type, which would be
otherwise compatible with the type of the identifiers, a method explosion
seemed inevitable.

With the new Mappable interface, it was just enough to bring a new method
PublicSymbol::mappable that returns a Mappable instance which uses the
PublicSymbol::identifier method as the mappable key supplier.

There is yet to be explored whether collections could employ the new
interface instead of Symbol.