Let collections use Mappable instead of Symbol
The generalization seems possible. Just each place using the key must use
now Mappable::mappable to get the correct key instead of using the symbol
instance directly.

As a benefit, it was possible to refine the signature of computing methods
that get the key; now the key type is finer.