Rename 'gain' and 'seek' methods
Renaming 'gain' to 'getOrDefault' and 'seek' to 'findOrDefault'. The
methods have the semantics that is better expressed by the new names,
which are also better consistent with Map::getOrDefault.

The major reason for the original method names was their (short) length,
yet being synonymous to the related methods. However, the names were too
confusing; the new names express the relationship better.