Overhaul the adaptation core
When testing implementation of more specific symbols, it turned out that
the number of convenient methods grows too much and copy&paste becomes
tempting. The major cause of the problem is the existence of fallbacks.
While fallbacks are useful, working with them is not so easy sometimes:
some cases need just the adaptation result, some cases need fallback if
the adaptation fails, some require a non-null value etc.

To avoid the emerging method explosion, just the most typical methods
should be provided directly by the AdaptationStrategy, the remaining cases
should be covered with a helper result type, which resembles the Optional.
All the other cases can then exploit the single implementation, especially
when it allows chaining to support all fallback-related combinations.

Introduction of the new AdaptationResult type resulted in a cascade of
reductions and modifications, starting with the Mappable interface and
stopping at collections. A new interface had to be created in order to
allow AdaptationStrategy to become the adaptation point handling all the
fallback problems. The new interface, AdaptationProvider, focused to the
very implementation of the adaptation, so that actual implementations do
not have to deal with other details. It became clear where referring to
the pure implementation is meant as well.