Revamp toString support
The mixture of introspection and toString was not quite right. Now,
toString uses introspection, but does not require to make anything special
for passing an introspective attribute to the result. Instead, the methods
were redesigned and simplified, so that the rendering can be customized in
any way if the default (which renders all attributes) is not satisfying.

Moreover, toString implementation became caching the result of the
rendering, so that more complex rendering (which the default is) does not
make too large performance penalty. The assumption is that toString might
be invoked quite often, especially for logging purposes (especially when
acting as keys in maps that are dumped in a log).

However, to keep the implementation choice possible, the base class for
symbols is now a new class: SymbolCore. AbstractSymbol extends it with the
introspection and toString support, keeping the original hierarchy.